Car Information System
So basically, you make a file such as the example one in
vehicles\vehicleinfo\ , and VehicleType can read the data in. You can
replace this with XML if you want, but we are VERY short on time, so
that won't work. File reading works, but isn't connected to anything
yet. Next commit will be deleting random stuff.